+++ src/main/java/net/mcreator/scpadditions/init/ScpAdditionsModBlockEntities.java
@@ -15,11 +15,13 @@ import net.mcreator.scpadditions.block.entity.Scp294StockingBlockEntity;
 import net.mcreator.scpadditions.block.entity.Scp294OutOfRangeBlockEntity;
 import net.mcreator.scpadditions.block.entity.Scp294BlockEntity;
 import net.mcreator.scpadditions.block.entity.Scp330BlockEntity;
+import net.mcreator.scpadditions.block.entity.SystemTerminalBlockEntity;
 import net.mcreator.scpadditions.ScpAdditionsMod;
 
 public class ScpAdditionsModBlockEntities {
 	public static final DeferredRegister<BlockEntityType<?>> REGISTRY = DeferredRegister.create(ForgeRegistries.BLOCK_ENTITY_TYPES, ScpAdditionsMod.MODID);
 	public static final RegistryObject<BlockEntityType<Scp330BlockEntity>> SCP_330 = REGISTRY.register("scp_330", () -> BlockEntityType.Builder.of(Scp330BlockEntity::new, ScpAdditionsModBlocks.SCP_330.get()).build(null));
+	public static final RegistryObject<BlockEntityType<SystemTerminalBlockEntity>> SCP_079_SYSTEM_CONTROL = REGISTRY.register("scp_079_system_control", () -> BlockEntityType.Builder.of(SystemTerminalBlockEntity::new, ScpAdditionsModBlocks.SCP_079_SYSTEM_CONTROL.get()).build(null));
 	public static final RegistryObject<BlockEntityType<?>> SCP_294 = register("scp_294", ScpAdditionsModBlocks.SCP_294, Scp294BlockEntity::new);
 	public static final RegistryObject<BlockEntityType<?>> SCP_294_OUT_OF_RANGE = register("scp_294_out_of_range", ScpAdditionsModBlocks.SCP_294_OUT_OF_RANGE, Scp294OutOfRangeBlockEntity::new);
 	public static final RegistryObject<BlockEntityType<?>> SCP_294_STOCKING = register("scp_294_stocking", ScpAdditionsModBlocks.SCP_294_STOCKING, Scp294StockingBlockEntity::new);